Work Template Code Refectoring done

+++ DTOs/WorkPlanning/WorkPlanSegmentDto.cs
@@ -4,7 +4,7 @@
     {
         public int WorkPlanSegmentId { get; set; }
 
-        public int WorkPlanId { get; set; }
+        public long WorkPlanId { get; set; }
 
         public int? WorkTemplateSegmentId { get; set; }
